+++ SonicGenerationsAP/Utils.hpp
@@ -1,8 +1,8 @@
 namespace Utils
 {
 	template<class T>
-	T* Offset(LPCWSTR lpModuleName, uintptr_t offset)
+	T* Offset(LPCWSTR lpModuleName, uint32_t offset)
 	{
-		return reinterpret_cast<T*>(reinterpret_cast<uintptr_t>(GetModuleHandle(lpModuleName)) + offset);
+		return reinterpret_cast<T*>(reinterpret_cast<uint32_t>(GetModuleHandle(lpModuleName)) + offset);
 	}
 }
